Move RunnerDetails to its own file
