+++ HomeRunTracker.Common/Models/Details/MlbPlayRunner.cs
@@ -13,19 +13,3 @@ public class MlbPlayRunner
     [Id(1)]
     public RunnerDetails Details { get; set; } = new();
 }
-
-[GenerateSerializer]
-public class RunnerDetails
-{
-    [JsonProperty("responsiblePitcher")]
-    [Id(0)]
-    public Pitcher Pitcher { get; set; } = new();
-    
-    [JsonProperty("isScoringEvent")]
-    [Id(1)]
-    public bool IsScoringEvent { get; set; }
-    
-    [JsonProperty("earned")]
-    [Id(2)]
-    public bool Earned { get; set; }
-}
